Rating/Commenting Functionality
Also includes changes made to models.py and schemas.py to accommodate

Note: response_model=List[BookSchema]
causes an error when I test the book browsing functionality (It works fine without it). I assume it's a result of my changes to models.py and/or schema.py but I'm not knowledgeable enough to know how to fix it.